Small refactors and the terrible test level!
Finally included the terrible temp.dat level file, so if you want to try this out without starting up the level editor, you can!
Messed about with the constats file, added new constants to aid later refactoring
Refactored loading animations, it should be quite a simple process now: Just populate loadedAnimations in the derived-from-Entity class, and your animations are cloned and set-up automagically by the superclass!
A note on the above: Filling loadedAnimations should be done on either start-up or during a level load.
Load directly to game, I'm sick of the title screen!

TODO: Extend the animation stuff to the tile framework, maybe merging the whole thing with the entity framework.